+++ CHANGELOG.md
@@ -7,12 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2]
+
+### Added
+
+- Add support for `authenticator_groups_config`
+
+## Changed
+
+- BREAKING: Upgrade provider to `~4.16` to add support for `authenticator_groups_config`
+
 ## [0.0.1]
 
 ### Added
 
 - Add support for `google_container_cluster` with autopilot enabled
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-gke-autopilot-cluster/compare/v0.0.1...HEAD
-<!-- [0.0.2]: https://github.com/mineiros-io/terraform-google-gke-autopilot-cluster/compare/v0.0.1...v0.0.2 -->
+[unreleased]: https://github.com/mineiros-io/terraform-google-gke-autopilot-cluster/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/mineiros-io/terraform-google-gke-autopilot-cluster/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/mineiros-io/terraform-google-gke-autopilot-cluster/releases/tag/v0.0.1